menage

+++ src/Chantier.cpp
@@ -2,7 +2,7 @@
 
 void Chantier::retirerTuile(int id)
 {
-    if (id >= 0 && id < getTaille())
+    if (id >= 0 && static_cast<size_t>(id) < getTaille())
     {
         tuilesChantier.erase(tuilesChantier.begin() + id);
     }
